moving helpers functions to utils and updating readme
