+++ src/utils/validateIncrement/index.js.flow
